+++ pandapower/convert_format.py
@@ -67,10 +67,31 @@ def _convert_q_capability_characteristic(net: pandapowerNet):
     # this is necessary due to the fact that Excel sheet names have a limit of 31 characters
     if 'q_capability_curve_characteristic' in net:
         net['q_capability_characteristic'] = net.pop('q_capability_curve_characteristic')
+    if 'q_capability_characteristic' not in net:
+        net['q_capability_characteristic'] = pd.DataFrame({
+            'id_q_capability_curve': pd.Series(dtype='Int64'),
+            'q_max_characteristic': pd.Series(dtype='object'),
+            'q_min_characteristic': pd.Series(dtype='object')
+        })
+    if 'q_capability_curve_table' not in net:
+        net['q_capability_curve_table'] = pd.DataFrame({
+            'id_q_capability_curve': pd.Series(dtype='Int64'),
+            'p_mw': pd.Series(dtype='float64'),
+            'q_min_mvar': pd.Series(dtype='float64')
+        })
     for ele in ['gen', 'sgen']:
         if isinstance(net[ele], pd.DataFrame) and 'id_q_capability_curve_characteristic' in net[ele].columns:
             net[ele] = net[ele].rename(
                 columns={'id_q_capability_curve_characteristic': 'id_q_capability_characteristic'})
+        if isinstance(net[ele], pd.DataFrame) and 'id_q_capability_characteristic' not in net[ele].columns:
+            net[ele]['id_q_capability_characteristic'] = pd.Series(dtype='Int64')
+        if isinstance(net[ele], pd.DataFrame) and 'reactive_capability_curve' not in net[ele].columns:
+            net[ele]['reactive_capability_curve'] = False
+    # add shunt information if not added yet
+    if isinstance(net.shunt, pd.DataFrame) and 'id_characteristic_table' not in net.shunt.columns:
+        net.shunt['id_characteristic_table'] = pd.Series(dtype='Int64')
+    if isinstance(net.shunt, pd.DataFrame) and 'step_dependency_table' not in net.shunt.columns:
+        net.shunt['step_dependency_table'] = False
 
 
 def _convert_geo_data(net, elements_to_deserialize=None):